Update scf.rst
According to {Besalú, E., Carbó-Dorca, R. The general Gaussian product theorem. J Math Chem 49, 1769–1784 (2011)}, the R^2 should be over the fraction line which also leads for correct results for me (which makes sense since otherwise the function would be sometimes not defined). I guess it was already meant like this in the script but for some of us the original notation was rather misleading. 
I hope the formatting is still fine, otherwise I would be thankful if you could correct it.